Major refactoring and added visualization effects
Refactoring game logic to use lists instead of mapping when possibe when index is player id. Refactoring cage to have less timeouts on the event queue so that it is easier to manage and follow. Adding visualizations for mouse events. Creating mouseAvatar file to clean up visualization logic. Adding mouse sprite that indicates sex and also moves tail when turning to add to quality of user experience. Adding status bars at the bottom of visualization to show how player is doing.