This tries to deliver a german about.html
In the german 'branch' of blockly games is an completely english about.html.

In the GIT master of blockly-games there is no different place for a TRANSLATED about.html.
Probably it has to be created somehow.

But I think that rather not the majority of young kids in germany speak english. So a translated about.html
could be usefull.